+++ examples/example/react/src/router.tsx
@@ -6,6 +6,7 @@ const Home = lazy(() => import('./pages/Home'))
 
 // Basic
 const ButtonDemo = lazy(() => import('./pages/ButtonDemo'))
+const SplitButtonDemo = lazy(() => import('./pages/SplitButtonDemo'))
 const ConfigProviderDemo = lazy(() => import('./pages/ConfigProviderDemo'))
 const IconDemo = lazy(() => import('./pages/IconDemo'))
 const LinkDemo = lazy(() => import('./pages/LinkDemo'))
@@ -161,6 +162,7 @@ const router = createHashRouter([
       { index: true, element: <Home /> },
       // Basic
       { path: 'button', element: <ButtonDemo /> },
+      { path: 'split-button', element: <SplitButtonDemo /> },
       { path: 'config-provider', element: <ConfigProviderDemo /> },
       { path: 'icon', element: <IconDemo /> },
       { path: 'link', element: <LinkDemo /> },